+++ src/main/java/com/nvidia/spark/rapids/jni/kudo/SliceInfo.java
@@ -17,11 +17,11 @@
 package com.nvidia.spark.rapids.jni.kudo;
 
 public class SliceInfo {
-  final long offset;
-  final long rowCount;
+  final int offset;
+  final int rowCount;
   private final SlicedValidityBufferInfo validityBufferInfo;
 
-  SliceInfo(long offset, long rowCount) {
+  SliceInfo(int offset, int rowCount) {
     this.offset = offset;
     this.rowCount = rowCount;
     this.validityBufferInfo = SlicedValidityBufferInfo.calc(offset, rowCount);
@@ -31,11 +31,11 @@ public class SliceInfo {
     return validityBufferInfo;
   }
 
-  public long getOffset() {
+  public int getOffset() {
     return offset;
   }
 
-  public long getRowCount() {
+  public int getRowCount() {
     return rowCount;
   }
