+++ src/main/java/com/nvidia/spark/rapids/jni/kudo/SlicedValidityBufferInfo.java
@@ -20,13 +20,13 @@ package com.nvidia.spark.rapids.jni.kudo;
  * A simple utility class to hold information about serializing/deserializing sliced validity buffer.
  */
 class SlicedValidityBufferInfo {
-  private final long bufferOffset;
-  private final long bufferLength;
+  private final int bufferOffset;
+  private final int bufferLength;
   /// The bit offset within the buffer where the slice starts
   private final long beginBit;
   private final long endBit; // Exclusive
 
-  SlicedValidityBufferInfo(long bufferOffset, long bufferLength, long beginBit, long endBit) {
+  SlicedValidityBufferInfo(int bufferOffset, int bufferLength, long beginBit, long endBit) {
     this.bufferOffset = bufferOffset;
     this.bufferLength = bufferLength;
     this.beginBit = beginBit;
@@ -39,11 +39,11 @@ class SlicedValidityBufferInfo {
         ", beginBit=" + beginBit + ", endBit=" + endBit + '}';
   }
 
-  public long getBufferOffset() {
+  public int getBufferOffset() {
     return bufferOffset;
   }
 
-  public long getBufferLength() {
+  public int getBufferLength() {
     return bufferLength;
   }
 
@@ -55,16 +55,16 @@ class SlicedValidityBufferInfo {
     return endBit;
   }
 
-  static SlicedValidityBufferInfo calc(long rowOffset, long numRows) {
+  static SlicedValidityBufferInfo calc(int rowOffset, int numRows) {
     if (rowOffset < 0) {
       throw new IllegalArgumentException("rowOffset must be >= 0, but was " + rowOffset);
     }
     if (numRows < 0) {
       throw new IllegalArgumentException("numRows must be >= 0, but was " + numRows);
     }
-    long bufferOffset = rowOffset / 8;
+    int bufferOffset = rowOffset / 8;
     long beginBit = rowOffset % 8;
-    long bufferLength = 0;
+    int bufferLength = 0;
     if (numRows > 0) {
       bufferLength = (rowOffset + numRows - 1) / 8 - bufferOffset + 1;
     }
